add support for sessions
In order to keep track of activity in a stateless world, a cookie with a
session id needs to be set.  That session id can be used to lookup state for a
particular user using in-memory or a shared cache.  Different strategies and
an adapter pattern still need to be built.